fix(gtfs): correct type hints for mypy compliance
The mypy static type checker reported an "incompatible default"
error for the 'files' parameter, as its type was `list[str]` but
its default value was `None`.

This commit updates the type hint to `Optional[list[str]]` to
explicitly allow `None` as a value. It also corrects the
type for the `dtype` parameter to `Any` to accurately reflect
its accepted types.